Remove unnecessary join calls
A couple spots used `Control.Monad.join` in ways that were
more confusing than useful. In particular, they used STM
to calculate an IO action, when in fact they simply needed
to run an STM action and then use the result more directly.
Not a big deal, but this way seems clearer.